sun50i: Add basic platform support
This platform supports Allwinner's sun50iwXpX SoCs, including the A64,
H5, and H64. They include a single cluster of four Cortex-A53 cores.

Currently, only building BL31 is supported. It is expected to be used in
conjunction with a U-Boot SPL for DRAM initialzation and loading BL33.

In its current state, the platform will boot into U-Boot and Linux with
a single core. It is also able to cleanly reboot.

Signed-off-by: Samuel Holland <samuel@sholland.org>