allwinner: Implement full PSCI with suspend using SCPI
If an SCP firmware is present and able to communicate via SCPI, then use
that to implement CPU and system power state transitions, including CPU
and system suspend. Otherwise, fall back to the existing "native" CPU
power control implementation.

The SCPI shared memory is at the very end of the area reserved for the
SCP firmware.

Signed-off-by: Samuel Holland <samuel@sholland.org>