allwinner: Implement full PSCI with suspend using SCPI
If an SCP firmware is present and able to communicate via SCPI, then use
that to implement CPU and system power state transitions, including CPU
and system suspend. Otherwise, fall back to the existing "native" CPU
power control implementation.

The SCPI shared memory is at the very end of SRAM A2. 16KiB, including
the shared memory area, is reserved for the SCP firmware.

Because the SCP firmware is not loaded adjacent to the ARISC exception
vector area, the jump instructions cannot be included in the SCP
firmware image, and must be initialized here before turning on the SCP.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Change-Id: I37b9b9636f94d4125230423726f3ac5e9cdb551c